fix: failing tests which were not aligning with new logic
